move url normalize funcs
